format and change int32 to usize

+++ hoomd-spatial/src/lib.rs
@@ -9,9 +9,8 @@
 )]
 /*! TODO DOCS */
 
-use std::collections::HashMap;
 use hoomd_vector::Cartesian;
-
+use std::collections::HashMap;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum ParticleFlag {
@@ -19,19 +18,18 @@ pub enum ParticleFlag {
     Ghost,
 }
 
-
 pub struct CellList<const N: usize> {
     pub cell_width: f64,
-    pub cell_idx_to_particle_indices: HashMap<[i32; N], Vec<i32>>,
-    pub particle_idx_to_cell_index: HashMap<i32, [i32; N]>,
-    pub particle_max_index: i32,
+    pub cell_idx_to_particle_indices: HashMap<[usize; N], Vec<usize>>,
+    pub particle_idx_to_cell_index: HashMap<usize, [usize; N]>,
+    pub particle_max_index: usize,
 }
 
-impl<const N:usize> CellList<N> {
-
+// TODO change usize to usize
+impl<const N: usize> CellList<N> {
     #[inline]
-    fn cell_index_from_position(cell_width: f64, position: &Cartesian<N>) -> [i32; N] {
-        std::array::from_fn(|j| (position.coordinates[j] /cell_width).floor() as i32)
+    fn cell_index_from_position(cell_width: f64, position: &Cartesian<N>) -> [usize; N] {
+        std::array::from_fn(|j| (position.coordinates[j] / cell_width).floor() as usize)
     }
 
     pub fn new(cell_width: f64, positions: &Vec<Cartesian<N>>) -> Self {
@@ -41,7 +39,7 @@ impl<const N:usize> CellList<N> {
             // Create the cell index for each particle.
             // Use add_particle fn
             let cell_idx = Self::cell_index_from_position(cell_width, position);
-            let particle_index: i32 = i as i32;
+            let particle_index: usize = i as usize;
             cell_idx_to_particle_indices
                 .entry(cell_idx)
                 .or_insert(Vec::new())
@@ -52,11 +50,11 @@ impl<const N:usize> CellList<N> {
             cell_width,
             cell_idx_to_particle_indices,
             particle_idx_to_cell_index,
-            particle_max_index: positions.len() as i32,
+            particle_max_index: positions.len() as usize,
         }
     }
 
-    pub fn cell_index_from_particle_index(&self, particle_index: i32) -> Option<&[i32; N]> {
+    pub fn cell_index_from_particle_index(&self, particle_index: usize) -> Option<&[usize; N]> {
         self.particle_idx_to_cell_index.get(&particle_index)
     }
 
@@ -67,24 +65,38 @@ impl<const N:usize> CellList<N> {
             .entry(cell_idx)
             .or_insert(Vec::new())
             .push(particle_index);
-        self.particle_idx_to_cell_index.insert(particle_index, cell_idx);
+        self.particle_idx_to_cell_index
+            .insert(particle_index, cell_idx);
     }
 
-    pub fn remove_particle(&mut self, particle_index: i32) {
+    pub fn remove_particle(&mut self, particle_index: usize) {
         //  unwrap will panic if particle index is not present
         // TODO think about using entry_and_modify to make this code cleaner
-        if let Some(cell_idx) = self.particle_idx_to_cell_index.remove(&particle_index){
-            let particle_indices = self.cell_idx_to_particle_indices.get_mut(&cell_idx).expect("Cell index found in the cell list.");
-            let index = particle_indices.iter().position(|&x| x == particle_index).expect("Particle index is in the cell list.");
+        if let Some(cell_idx) = self.particle_idx_to_cell_index.remove(&particle_index) {
+            let particle_indices = self
+                .cell_idx_to_particle_indices
+                .get_mut(&cell_idx)
+                .expect("Cell index found in the cell list.");
+            let index = particle_indices
+                .iter()
+                .position(|&x| x == particle_index)
+                .expect("Particle index is in the cell list.");
             particle_indices.swap_remove(index);
         } else {
             panic!("Particle index not found in cell list");
         }
     }
 
-    pub fn translate_particle(&mut self, particle_index: i32, new_particle_position: Cartesian<N>){
-        let current_cell_idx = self.particle_idx_to_cell_index.get(&particle_index).expect("Particle index found in the cell list.");
-        let new_cell_idx = Self::cell_index_from_position(self.cell_width,&new_particle_position);
+    pub fn translate_particle(
+        &mut self,
+        particle_index: usize,
+        new_particle_position: Cartesian<N>,
+    ) {
+        let current_cell_idx = self
+            .particle_idx_to_cell_index
+            .get(&particle_index)
+            .expect("Particle index found in the cell list.");
+        let new_cell_idx = Self::cell_index_from_position(self.cell_width, &new_particle_position);
         if *current_cell_idx != new_cell_idx {
             self.remove_particle(particle_index);
             self.add_particle(&new_particle_position);
